--- app/db/db.js
@@ -1,22 +1,0 @@
-const orders = [
-  {
-    orderId: 1,
-    foodName: 'chicken and chips',
-    price: 'N2000',
-    quantity: '5',
-    orderedBy: 'Robert Downey',
-    orderDatetime: '24-08-2018 20:00:09',
-    orderStatus: 'Pending',
-  },
-  {
-    orderId: 2,
-    foodName: 'Shawarma and sauce',
-    price: 'N5000',
-    quantity: '5',
-    orderedBy: 'Dr Strange',
-    orderDatetime: '24-08-2018 20:00:09',
-    orderStatus: 'Pending',
-  },
-];
-
-export default orders;
