+++ app/db/orders.db.js
@@ -1,0 +1,24 @@
+const orders = [
+  {
+    orderId: 1,
+    foodId: 2,
+    userId: 3,
+    quantity: '5',
+    amount: 'N2000',
+    orderStatus: 'Pending',
+    orderDatetime: '24-08-2018 20:00:09',
+    updatedAt: '24-08-2018 20:00:09',
+  },
+  {
+    orderId: 2,
+    foodId: 2,
+    userId: 4,
+    quantity: '5',
+    amount: 'N2000',
+    orderStatus: 'Pending',
+    orderDatetime: '24-08-2018 20:00:09',
+    updatedAt: '24-08-2018 20:00:09',
+  },
+];
+
+export default orders;
